+++ continual/container.py
@@ -616,7 +616,7 @@ def Residual(
     module: CoModule,
     temporal_fill: PaddingMode = None,
     reduce: Reduction = "sum",
-    phantom_padding: bool = False,
+    residual_shrink: bool = False,
 ) -> BroadcastReduce:
     """[summary]
 
@@ -624,7 +624,7 @@ def Residual(
         module (CoModule): module to which a residual should be added.
         temporal_fill (PaddingMode, optional): temporal fill type in delay. Defaults to None.
         reduce (Reduction, optional): Reduction function. Defaults to "sum".
-        phantom_padding (bool, optional):
+        residual_shrink (bool, optional):
             Set residual delay to operate as if equal padding was used for the . Defaults to False.
 
     Returns:
@@ -638,15 +638,15 @@ def Residual(
         module, "temporal_fill", PaddingMode.REPLICATE.value
     )
     equal_padding = module.receptive_field - num_from(module.padding) * 2 == 1
-    do_phantom_padding = phantom_padding and not equal_padding
+    do_residual_shrink = residual_shrink and not equal_padding
 
     delay = module.delay
-    if do_phantom_padding:
+    if do_residual_shrink:
         assert delay % 2 == 0, "Auto-shrink only works for even-number delays."
         delay = delay // 2
     return BroadcastReduce(
         # Residual first yields easier broadcasting in reduce functions
-        Delay(delay, temporal_fill, auto_shrink=do_phantom_padding),
+        Delay(delay, temporal_fill, auto_shrink=do_residual_shrink),
         module,
         reduce=reduce,
         auto_delay=False,