+++ continual/__init__.py
@@ -27,6 +27,6 @@ from .pooling import (  # noqa: F401
     MaxPool3d,
 )
 from .ptflops import _register_ptflops  # noqa: F401
-from .rnn import RNN  # noqa: F401
+from .rnn import LSTM, RNN  # noqa: F401
 from .shape import Reshape  # noqa: F401
 from .utils import flat_state_dict, load_state_dict, state_dict  # noqa: F401