+++ CHANGELOG.md
@@ -13,6 +13,7 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 - "lagging" option for shrink in `co.Delay` and `co.Residual`.
 - `co.RNN`.
 - `co.LSTM`.
+- `co.GRU`.
 
 ### Changed
 - `phantom_padding` renamed to `residual_shrink`.